application/backend/requirements.txt
fastapi==0.116.1
SQLAlchemy==2.0.43
passlib==1.7.4
uvicorn==0.35.0
bcrypt==4.3.0
passlib==1.7.4
python-jose==3.5.0
python-multipart==0.0.20
python-dotenv